+++ src/main/java/codesquad/db/csv/CsvDataSourceManager.java
@@ -1,10 +1,9 @@
-package codesquad.db;
+package codesquad.db.csv;
 
 import codesquad.container.Component;
+import codesquad.db.ConnectionPoolManager;
+import codesquad.db.DataSourceConfigurer;
 import codesquad.file.AppFileReader;
-import org.h2.jdbcx.JdbcConnectionPool;
-import org.h2.jdbcx.JdbcDataSource;
-import org.h2.tools.Server;
 
 import javax.sql.DataSource;
 import java.io.File;
@@ -14,33 +13,24 @@ import java.sql.DriverManager;
 import java.sql.SQLException;
 import java.sql.Statement;
 
-public class H2ConnectionPoolManager implements ConnectionPoolManager {
+@Component
+public class CsvDataSourceManager implements ConnectionPoolManager {
     private static final String homePath = System.getProperty("user.home");
     private static final String CREATE_SQL_FILE_PATH = "db/create.sql";
     private static final String DROP_SQL_FILE_PATH = "db/drop.sql";
     private File DB_DIR;
 
-    private final JdbcConnectionPool jdbcConnectionPool;
 
-
-    public H2ConnectionPoolManager(DataSourceConfigurer dataSourceConfigurer) {
+    public CsvDataSourceManager(DataSourceConfigurer dataSourceConfigurer) {
         try {
             initFile();
-            initServer();
             initTable(dataSourceConfigurer);
-        } catch (SQLException | IOException e) {
+        } catch (SQLException e) {
+            throw new RuntimeException(e);
+        } catch (IOException e) {
             throw new RuntimeException(e);
         }
 
-        JdbcDataSource jdbcDataSource = new JdbcDataSource();
-        jdbcDataSource.setURL(dataSourceConfigurer.getURL());
-        jdbcDataSource.setUser(dataSourceConfigurer.getUsername());
-        jdbcDataSource.setPassword(dataSourceConfigurer.getPassword());
-        this.jdbcConnectionPool = JdbcConnectionPool.create(jdbcDataSource);
-    }
-
-    private void initServer() throws SQLException {
-        Server.createTcpServer("-tcp", "-tcpPort", "9092").start();
     }
 
     private void initFile() throws IOException {
@@ -49,16 +39,27 @@ public class H2ConnectionPoolManager implements ConnectionPoolManager {
             wasDir.mkdirs();
         }
 
-        final String UPLOAD_DIR_PATH = "db";
+        final String UPLOAD_DIR_PATH = "csv";
         DB_DIR = new File(wasDir, UPLOAD_DIR_PATH);
 
         if (!DB_DIR.exists()) {
             DB_DIR.mkdirs();
         }
 
-        File wasDb = new File(DB_DIR, "was.mv.db");
-        if (!wasDb.exists()) {
-            wasDb.createNewFile();
+
+        File users = new File(DB_DIR, "users.csv");
+        if (!users.exists()) {
+            users.createNewFile();
+        }
+
+        File posts = new File(DB_DIR, "posts.csv");
+        if (!posts.exists()) {
+            posts.createNewFile();
+        }
+
+        File comments = new File(DB_DIR, "comments.csv");
+        if (!comments.exists()) {
+            comments.createNewFile();
         }
     }
 
@@ -69,7 +70,6 @@ public class H2ConnectionPoolManager implements ConnectionPoolManager {
                 dataSourceConfigurer.getPassword()
         );
 
-
         Statement statement = connection.createStatement();
 
         AppFileReader dropSql = new AppFileReader(DROP_SQL_FILE_PATH);
@@ -78,14 +78,21 @@ public class H2ConnectionPoolManager implements ConnectionPoolManager {
         switch (dataSourceConfigurer.ddlOption()) {
             case DROP_CREATE -> {
                 statement.execute(dropSql.getContent());
-                statement.execute(createSql.getContent());
+                executeCreate(statement, createSql.getContent());
             }
-            case CREATE -> statement.execute(createSql.getContent());
+            case CREATE -> executeCreate(statement, createSql.getContent());
+        }
+    }
+
+    private void executeCreate(Statement statement, String sqls) throws SQLException {
+        String[] splitSqls = sqls.split(";");
+        for (String sql : splitSqls) {
+            statement.execute(sql.trim());
         }
     }
 
     @Override
     public DataSource getDataSource() {
-        return jdbcConnectionPool;
+        return new CsvDatasource();
     }
 }
