+++ src/main/java/codesquad/db/h2/H2ConnectionPoolManager.java
@@ -1,8 +1,11 @@
-package codesquad.db;
+package codesquad.db.h2;
 
-import codesquad.container.Component;
-import codesquad.db.csv.CsvDatasource;
+import codesquad.db.ConnectionPoolManager;
+import codesquad.db.DataSourceConfigurer;
 import codesquad.file.AppFileReader;
+import org.h2.jdbcx.JdbcConnectionPool;
+import org.h2.jdbcx.JdbcDataSource;
+import org.h2.tools.Server;
 
 import javax.sql.DataSource;
 import java.io.File;
@@ -12,24 +15,33 @@ import java.sql.DriverManager;
 import java.sql.SQLException;
 import java.sql.Statement;
 
-@Component
-public class CsvDataSourceManager implements ConnectionPoolManager {
+public class H2ConnectionPoolManager implements ConnectionPoolManager {
     private static final String homePath = System.getProperty("user.home");
     private static final String CREATE_SQL_FILE_PATH = "db/create.sql";
     private static final String DROP_SQL_FILE_PATH = "db/drop.sql";
     private File DB_DIR;
 
+    private final JdbcConnectionPool jdbcConnectionPool;
 
-    public CsvDataSourceManager(DataSourceConfigurer dataSourceConfigurer) {
+
+    public H2ConnectionPoolManager(DataSourceConfigurer dataSourceConfigurer) {
         try {
             initFile();
+            initServer();
             initTable(dataSourceConfigurer);
-        } catch (SQLException e) {
-            throw new RuntimeException(e);
-        } catch (IOException e) {
+        } catch (SQLException | IOException e) {
             throw new RuntimeException(e);
         }
 
+        JdbcDataSource jdbcDataSource = new JdbcDataSource();
+        jdbcDataSource.setURL(dataSourceConfigurer.getURL());
+        jdbcDataSource.setUser(dataSourceConfigurer.getUsername());
+        jdbcDataSource.setPassword(dataSourceConfigurer.getPassword());
+        this.jdbcConnectionPool = JdbcConnectionPool.create(jdbcDataSource);
+    }
+
+    private void initServer() throws SQLException {
+        Server.createTcpServer("-tcp", "-tcpPort", "9092").start();
     }
 
     private void initFile() throws IOException {
@@ -38,27 +50,16 @@ public class CsvDataSourceManager implements ConnectionPoolManager {
             wasDir.mkdirs();
         }
 
-        final String UPLOAD_DIR_PATH = "csv";
+        final String UPLOAD_DIR_PATH = "db";
         DB_DIR = new File(wasDir, UPLOAD_DIR_PATH);
 
         if (!DB_DIR.exists()) {
             DB_DIR.mkdirs();
         }
 
-
-        File users = new File(DB_DIR, "users.csv");
-        if (!users.exists()) {
-            users.createNewFile();
-        }
-
-        File posts = new File(DB_DIR, "posts.csv");
-        if (!posts.exists()) {
-            posts.createNewFile();
-        }
-
-        File comments = new File(DB_DIR, "comments.csv");
-        if (!comments.exists()) {
-            comments.createNewFile();
+        File wasDb = new File(DB_DIR, "was.mv.db");
+        if (!wasDb.exists()) {
+            wasDb.createNewFile();
         }
     }
 
@@ -69,6 +70,7 @@ public class CsvDataSourceManager implements ConnectionPoolManager {
                 dataSourceConfigurer.getPassword()
         );
 
+
         Statement statement = connection.createStatement();
 
         AppFileReader dropSql = new AppFileReader(DROP_SQL_FILE_PATH);
@@ -77,21 +79,14 @@ public class CsvDataSourceManager implements ConnectionPoolManager {
         switch (dataSourceConfigurer.ddlOption()) {
             case DROP_CREATE -> {
                 statement.execute(dropSql.getContent());
-                executeCreate(statement, createSql.getContent());
+                statement.execute(createSql.getContent());
             }
-            case CREATE -> executeCreate(statement, createSql.getContent());
-        }
-    }
-
-    private void executeCreate(Statement statement, String sqls) throws SQLException {
-        String[] splitSqls = sqls.split(";");
-        for (String sql : splitSqls) {
-            statement.execute(sql.trim());
+            case CREATE -> statement.execute(createSql.getContent());
         }
     }
 
     @Override
     public DataSource getDataSource() {
-        return new CsvDatasource();
+        return jdbcConnectionPool;
     }
 }
