+++ src/test/java/codesquad/server/handler/RequestHandlerTest.java
@@ -34,8 +34,8 @@ class RequestHandlerTest {
                     "codesquad.server.handler.ErrorHandler",
                     "codesquad.http.security.SessionStorage",
                     "codesquad.app.storage.RdbmsUserDao",
-                    "codesquad.db.H2ConnectionPoolManager",
                     "codesquad.db.CsvDataSource",
+                    "codesquad.db.TestH2ConnectionPoolManager",
                     "codesquad.db.TestDataSourceConfigurer",
                     "codesquad.app.service.SessionService",
                     "codesquad.server.router.RouteEntryManager",
